fix navigating to the journal whose row is clicked in journal details page
