+++ src/renderer/hooks/index.ts
@@ -1,1 +1,2 @@
 export * from './useTheme';
+export * from './useAuth';
