+++ src/renderer/routes.tsx
@@ -1,14 +1,13 @@
+import { type PropsWithChildren } from 'react';
 import { MemoryRouter, Route, Routes, Navigate } from 'react-router-dom';
 
+import { ThemeProvider, AuthProvider, useAuth } from './hooks';
+import { Toaster } from './shad/ui/toaster';
+
+import Nav from './components/Nav';
 import Home from './views/Home';
 import Login from './views/Login';
 import Register from './views/Register';
-import { useContext } from 'react';
-import { AuthContext } from './context/Auth';
-import AuthProvider from './providers/Auth';
-import { ThemeProvider } from './hooks';
-import { Toaster } from './shad/ui/toaster';
-import Nav from './components/Nav';
 import AccountsPage from './views/Accounts';
 import LedgerPage from './views/Ledger';
 import JournalsPage from './views/Journals';
@@ -79,8 +78,8 @@ export default function appRoutes() {
   );
 }
 
-function RequireAuth({ children }: { children: JSX.Element }) {
-  let { authed } = useContext(AuthContext);
+function RequireAuth({ children }: PropsWithChildren) {
+  const { authed } = useAuth();
 
   if (authed || process.env.NODE_ENV === 'development') return children;
