+++ testing-bugs.txt
@@ -1,4 +1,3 @@
-// TODO: clicking on a specific journal on Journal page is not opening that journal
 // TODO: add setting option to set default label for when debit/credit is 0 => 0 / empty / dash (any character of length 1)
 // TODO: make narration optional in newJournal
 // TODO: add debounce effect in debit / credit inputs to enable/disable publish button on journal
